move the build_ui function of imag-ids into its own file
