+++ bin/core/imag-ids/src/main.rs
@@ -51,26 +51,6 @@ use libimagerror::io::ToExitCode;
 use libimagstore::storeid::StoreId;
 
 
-/// No special CLI
-pub fn build_ui<'a>(app: App<'a, 'a>) -> App<'a, 'a> {
-    app
-        .arg(Arg::with_name("print-storepath")
-             .long("with-storepath")
-             .takes_value(false)
-             .required(false)
-             .multiple(false)
-             .help("Print the storepath for each id"))
-
-        .arg(Arg::with_name("in-collection-filter")
-             .long("in-collection")
-             .short("c")
-             .required(false)
-             .takes_value(true)
-             .multiple(true)
-             .value_names(&["COLLECTION"])
-             .help("Filter for ids which are only in these collections"))
-}
-
 pub struct IsInCollectionsFilter<'a, A>(Option<A>, ::std::marker::PhantomData<&'a str>)
     where A: AsRef<[&'a str]>;
